StatsHandler -> StatsRequestHandler; moved to its own file
